+++ tradin/src/components/commons/breadcrumb/BreadCrumb.types.ts
@@ -1,0 +1,9 @@
+export interface BreadCrumbUIProps {
+    pathSegments: string[]
+    handleSave: (index: number) => void
+    handleEdit: (segment: string, index: number) => void
+    editingSegment: string | null
+    editValue: string
+    isEditablePath: (path: string[]) => boolean
+    setEditValue: React.Dispatch<React.SetStateAction<string>>
+}
